pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "phd-platform"
version = "0.1.0"
description = "PhD-level adaptive education platform sourcing Ivy League + Stanford/Berkeley/CMU curricula"
requires-python = ">=3.11"
license = {text = "MIT"}
dependencies = [
    "anthropic>=0.40.0",
    "pydantic>=2.5",
    "pydantic-settings>=2.2",
    "pyyaml>=6.0",
    "rich>=13.0",
    "httpx>=0.27",
    "polars>=1.0",
    "plotly>=5.18",
    "networkx>=3.2",
    "numpy>=1.26",
    "sqlalchemy[asyncio]>=2.0",
    "aiosqlite>=0.20",
    "alembic>=1.13",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-cov>=4.1",
    "ruff>=0.4",
    "mypy>=1.10",
]

[project.scripts]
phd-platform = "phd_platform.__main__:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.mypy]
python_version = "3.11"
strict = true
